+++ RadixWallet/Core/DesignSystem/Components/JaggedEdge.swift
@@ -62,9 +62,3 @@ struct JaggedEdgeShape: Shape {
 		}
 	}
 }
-
-extension CGRect {
-	subscript(unitPoint: UnitPoint) -> CGPoint {
-		.init(x: minX + unitPoint.x * width, y: minY + unitPoint.y * height)
-	}
-}
